add prompt before logging out

+++ src/index.jsx
@@ -1,9 +1,11 @@
 import React from "react";
 import { entrypoints, shell } from "uxp";
 
+import { ConfirmationDialog } from "./components/ConfirmationDialog";
+import { CommandController } from "./controllers/CommandController";
+import { PanelController } from "./controllers/PanelController";
+import { App } from "./panels/App";
 import "./styles.css";
-import { PanelController } from "./controllers/PanelController.jsx";
-import { App } from "./panels/App.jsx";
 
 const appPanelController = new PanelController(() => <App />, {
     id: "app",
@@ -13,9 +15,17 @@ const appPanelController = new PanelController(() => <App />, {
             label: "Logout",
             enabled: true,
             checked: false,
-            onInvoke: () => {
-                localStorage.clear();
-                location.reload();
+            onInvoke: async () => {
+                const confirmationDialogController = new CommandController(
+                    ({ dialog }) => <ConfirmationDialog dialog={dialog} />
+                );
+
+                const logout = await confirmationDialogController.run();
+
+                if (logout === "YES") {
+                    localStorage.clear();
+                    location.reload();
+                }
             },
         },
     ],
@@ -33,13 +43,19 @@ entrypoints.setup({
     },
     commands: {
         async goToDashboard() {
-            await shell.openExternal("https://console.pixelbin.io/choose-org?redirectTo=storage");
+            await shell.openExternal(
+                "https://console.pixelbin.io/choose-org?redirectTo=storage"
+            );
         },
         async buyCredits() {
-            await shell.openExternal("https://console.pixelbin.io/choose-org?redirectTo=settings/billing/pricing");
+            await shell.openExternal(
+                "https://console.pixelbin.io/choose-org?redirectTo=settings/billing/pricing"
+            );
         },
         async howItWorks() {
-            await shell.openExternal("https://www.pixelbin.io/docs/integrations/photoshop/");
+            await shell.openExternal(
+                "https://www.pixelbin.io/docs/integrations/photoshop/"
+            );
         },
     },
     panels: {
